// src/api/api-client/isi-backend/models/PersonalFilterRequestDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ISI Backend API
 * ISI Backend - Service für das Informationssystem für soziale Infrastrukturplanung
 *
 * The version of the OpenAPI document: 1.6.0
 * Contact: noreply@mail.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { FilterSettingsDto } from './FilterSettingsDto';
import {
    FilterSettingsDtoFromJSON,
    FilterSettingsDtoFromJSONTyped,
    FilterSettingsDtoToJSON,
    FilterSettingsDtoToJSONTyped,
} from './FilterSettingsDto';

/**
 * 
 * @export
 * @interface PersonalFilterRequestDto
 */
export interface PersonalFilterRequestDto {
    /**
     * 
     * @type {string}
     * @memberof PersonalFilterRequestDto
     */
    id?: string;
    /**
     * 
     * @type {FilterSettingsDto}
     * @memberof PersonalFilterRequestDto
     */
    filterSettings: FilterSettingsDto;
    /**
     * 
     * @type {string}
     * @memberof PersonalFilterRequestDto
     */
    filterName: string;
}

/**
 * Check if a given object implements the PersonalFilterRequestDto interface.
 */
export function instanceOfPersonalFilterRequestDto(value: object): value is PersonalFilterRequestDto {
    if (!('filterSettings' in value) || value['filterSettings'] === undefined) return false;
    if (!('filterName' in value) || value['filterName'] === undefined) return false;
    return true;
}

export function PersonalFilterRequestDtoFromJSON(json: any): PersonalFilterRequestDto {
    return PersonalFilterRequestDtoFromJSONTyped(json, false);
}

export function PersonalFilterRequestDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): PersonalFilterRequestDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'filterSettings': FilterSettingsDtoFromJSON(json['filterSettings']),
        'filterName': json['filterName'],
    };
}

export function PersonalFilterRequestDtoToJSON(json: any): PersonalFilterRequestDto {
    return PersonalFilterRequestDtoToJSONTyped(json, false);
}

export function PersonalFilterRequestDtoToJSONTyped(value?: PersonalFilterRequestDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'filterSettings': FilterSettingsDtoToJSON(value['filterSettings']),
        'filterName': value['filterName'],
    };
}